+++ src/full/Agda/TypeChecking/Errors/Names.hs
@@ -116,7 +116,9 @@ typeErrorName = \case
   ForcedConstructorNotInstantiated                           {} -> ForcedConstructorNotInstantiated_
   FunctionTypeInSizeUniv                                     {} -> FunctionTypeInSizeUniv_
   GeneralizationFailed                                       {} -> GeneralizationFailed_
-  GeneralizationPrepruneError                                {} -> GeneralizationPrepruneError_
+  GeneralizationPrepruneErrorRefinedContext                  {} -> GeneralizationPrepruneErrorRefinedContext_
+  GeneralizationPrepruneErrorCyclicDependencies              {} -> GeneralizationPrepruneErrorCyclicDependencies_
+  GeneralizationPrepruneErrorFailedToInstantiate             {} -> GeneralizationPrepruneErrorFailedToInstantiate_
   GeneralizeCyclicDependency                                 {} -> GeneralizeCyclicDependency_
   GeneralizeNotSupportedHere                                 {} -> GeneralizeNotSupportedHere_
   GeneralizedVarInLetOpenedModule                            {} -> GeneralizedVarInLetOpenedModule_